requirements.txt
Flask
Flask-MySQLdb
Flask-WTF
requests
urllib
mysql.connector